+++ kmock/src/commonMain/kotlin/tech/antibytes/kmock/verification/UnchainedAssertion.kt
@@ -8,11 +8,12 @@ package tech.antibytes.kmock.verification
 
 import co.touchlab.stately.collections.sharedMutableMapOf
 import tech.antibytes.kmock.KMockContract.Assertions
+import tech.antibytes.kmock.KMockContract.CloseableAssertionContext
 import tech.antibytes.kmock.KMockContract.Proxy
 
 internal class UnchainedAssertion(
     assertions: Assertions = Assertions
-) : BaseAssertionContext(assertions) {
+) : BaseAssertionContext(assertions), CloseableAssertionContext {
     private val frozenInvocations: MutableMap<Proxy<*, *>, Int> = sharedMutableMapOf()
     private val unfrozenInvocations: MutableMap<Proxy<*, *>, Int> = mutableMapOf()
 
@@ -23,6 +24,14 @@ internal class UnchainedAssertion(
             frozenInvocations
         } else {
             unfrozenInvocations
+        }
+    }
+
+    override fun Proxy<*, *>.hasNoFurtherInvocations() {
+        runAssertion(this) { callIndex ->
+            if (callIndex < calls) {
+                throw AssertionError("Only $callIndex of $calls invocations had been asserted.")
+            }
         }
     }