+++ kmock/src/commonTest/kotlin/tech/antibytes/kmock/verification/VerificationAndAssertionAsyncSpec.kt
@@ -59,6 +59,24 @@ class VerificationAndAssertionAsyncSpec {
     }
 
     @Test
+    @JsName("fn1a")
+    fun `Given asyncVerify is called it fails if the covered mock does not have the minimum amount of calls if a negative value was assigned`() = runBlockingTest {
+        // Given
+        val proxy = fixture.funProxyFixture()
+        val givenCalls = 1
+        val expectedCalls = 3
+
+        // When
+        val error = assertFailsWith<AssertionError> {
+            asyncVerify(atLeast = -expectedCalls) {
+                Expectation(proxy, fixture.listFixture(size = givenCalls))
+            }
+        }
+
+        error.message mustBe "Expected at least $expectedCalls calls, but found only $givenCalls."
+    }
+
+    @Test
     @JsName("fn2")
     fun `Given asyncVerify is called it fails if the covered mock does exceeds the maximum amount of calls`() = runBlockingTest {
         // Given
@@ -69,6 +87,24 @@ class VerificationAndAssertionAsyncSpec {
         // When
         val error = assertFailsWith<AssertionError> {
             asyncVerify(atMost = expectedCalls) {
+                Expectation(proxy, fixture.listFixture(size = givenCalls))
+            }
+        }
+
+        error.message mustBe "Expected at most $expectedCalls calls, but exceeded with $givenCalls."
+    }
+
+    @Test
+    @JsName("fn2a")
+    fun `Given asyncVerify is called it fails if the covered mock does exceeds the maximum amount of calls if a negative value was assigned`() = runBlockingTest {
+        // Given
+        val proxy = fixture.funProxyFixture()
+        val givenCalls = 3
+        val expectedCalls = 1
+
+        // When
+        val error = assertFailsWith<AssertionError> {
+            asyncVerify(atMost = -expectedCalls) {
                 Expectation(proxy, fixture.listFixture(size = givenCalls))
             }
         }
@@ -95,6 +131,24 @@ class VerificationAndAssertionAsyncSpec {
     }
 
     @Test
+    @JsName("fn3a")
+    fun `Given asyncVerify is called it fails if the covered mock does not have the exact minimum amount of calls with a negative number`() = runBlockingTest {
+        // Given
+        val proxy = fixture.funProxyFixture()
+        val givenCalls = 1
+        val expectedCalls = 3
+
+        // When
+        val error = assertFailsWith<AssertionError> {
+            asyncVerify(exactly = -expectedCalls, atLeast = 0) {
+                Expectation(proxy, fixture.listFixture(size = givenCalls))
+            }
+        }
+
+        error.message mustBe "Expected at least $expectedCalls calls, but found only $givenCalls."
+    }
+
+    @Test
     @JsName("fn4")
     fun `Given asyncVerify is called it fails if the covered mock does exceeds the exact maximum amount of calls`() = runBlockingTest {
         // Given
@@ -105,6 +159,24 @@ class VerificationAndAssertionAsyncSpec {
         // When
         val error = assertFailsWith<AssertionError> {
             asyncVerify(exactly = expectedCalls, atMost = 0) {
+                Expectation(proxy, fixture.listFixture(size = givenCalls))
+            }
+        }
+
+        error.message mustBe "Expected at most $expectedCalls calls, but exceeded with $givenCalls."
+    }
+
+    @Test
+    @JsName("fn4a")
+    fun `Given asyncVerify is called it fails if the covered mock does exceeds the exact maximum amount of calls with a negative number`() = runBlockingTest {
+        // Given
+        val proxy = fixture.funProxyFixture()
+        val givenCalls = 3
+        val expectedCalls = 1
+
+        // When
+        val error = assertFailsWith<AssertionError> {
+            asyncVerify(exactly = -expectedCalls, atMost = 0) {
                 Expectation(proxy, fixture.listFixture(size = givenCalls))
             }
         }